+++ CSharpConsoleSolution/DisplayApp/Program.cs
@@ -5,7 +5,12 @@ namespace DisplayApp
     {
         static void Main(string[] args)
         {
-            
+            //Reason in the REFLECTION.md file
+            Console.WriteLine("Cannot use the classes of the MathApp !");
+
+            //Console.WriteLine("Using MathApp classes inside DispayApp");
+            //MathematicalCalculation mathematicalCalculation = new MathematicalCalculation();
+            //mathematicalCalculation.PerformMathematicalCalculation();
         }
     }
 }
